Update
-change comparison_count variable from int to long for bigger data set
-fix timer bug
-add comparison count func for flash and merge sort

+++ Sorting Library/bubbleSort.cpp
@@ -16,7 +16,7 @@ void bubble_sort(int a[], int n) {
     }
 }
 
-void bubble_sort_with_counter(int a[], int n, int &comparison_count) {
+void bubble_sort_with_counter(int a[], int n, long long &comparison_count) {
     comparison_count = 0;   //Count the number of element comparisons.
     for (int i = 0; i < n - 1; i++) {
         for (int j = 0; j < n - i; j++) {
